ISS22 - refactor lib
