+++ modules/arrow/__init__.py
@@ -1,3 +1,3 @@
 from .contour import Contour
 from .imgcv2 import ImgCV2
-from .lib_arrow import get_centeroid
+from .lib import Arrow
